+++ src/resqml2_0_1/ContinuousProperty.cpp
@@ -120,7 +120,7 @@ std::string ContinuousProperty::getUomAsString() const
 	return gsoap_resqml2_0_1::soap_resqml20__ResqmlUom2s(gsoapProxy2_0_1->soap, uom);
 }
 
-EML2_NS::AbstractHdfProxy* ContinuousProperty::getValuesHdfProxyAndDatasetPathOfPatch(unsigned int patchIndex, std::string & datasetPath) const
+EML2_NS::AbstractHdfProxy* ContinuousProperty::getValuesHdfProxyAndDatasetPathOfPatch(uint64_t patchIndex, std::string & datasetPath) const
 {
 	if (patchIndex >= getPatchCount()) {
 		throw out_of_range("The values property patch is out of range");
@@ -146,7 +146,7 @@ double ContinuousProperty::getMaximumValue(uint64_t index) const
 	return prop->MaximumValue.size() <= index ? std::numeric_limits<double>::quiet_NaN() : prop->MaximumValue[index];
 }
 
-void ContinuousProperty::setMinimumValue(double value, unsigned int index) const
+void ContinuousProperty::setMinimumValue(double value, uint64_t index) const
 {
 	_resqml20__ContinuousProperty* prop = static_cast<_resqml20__ContinuousProperty*>(gsoapProxy2_0_1);
 
@@ -157,7 +157,7 @@ void ContinuousProperty::setMinimumValue(double value, unsigned int index) const
 	prop->MinimumValue[index] = value;
 }
 
-void ContinuousProperty::setMaximumValue(double value, unsigned int index) const
+void ContinuousProperty::setMaximumValue(double value, uint64_t index) const
 {
 	_resqml20__ContinuousProperty* prop = static_cast<_resqml20__ContinuousProperty*>(gsoapProxy2_0_1);
