+++ src/resqml2_0_1/Grid2dRepresentation.cpp
@@ -752,7 +752,7 @@ gsoap_resqml2_0_1::resqml20__PointGeometry* Grid2dRepresentation::createArray2dO
 }
 
 gsoap_resqml2_0_1::resqml20__PointGeometry* Grid2dRepresentation::createArray2dOfExplicitZ(
-	unsigned int patchIndex, double * zValues, EML2_NS::AbstractLocal3dCrs * localCrs,
+	uint64_t patchIndex, double * zValues, EML2_NS::AbstractLocal3dCrs * localCrs,
 	unsigned int numI, unsigned int numJ, EML2_NS::AbstractHdfProxy * proxy,
 	RESQML2_NS::Grid2dRepresentation * supportingRepresentation,
 	uint64_t startGlobalIndex,
@@ -819,7 +819,7 @@ gsoap_resqml2_0_1::resqml20__PointGeometry* Grid2dRepresentation::createArray2dO
 }
 
 gsoap_resqml2_0_1::resqml20__PointGeometry* Grid2dRepresentation::createArray2dOfExplicitZ(
-	unsigned int patchIndex, double * zValues, EML2_NS::AbstractLocal3dCrs * localCrs,
+	uint64_t patchIndex, double * zValues, EML2_NS::AbstractLocal3dCrs * localCrs,
 	unsigned int numI, unsigned int numJ, EML2_NS::AbstractHdfProxy * proxy,
 	double originX, double originY, double originZ,
 	double offsetIX, double offsetIY, double offsetIZ, double spacingI,
