+++ src/resqml2_0_1/PointsProperty.cpp
@@ -115,7 +115,7 @@ COMMON_NS::DataObjectReference PointsProperty::getHdfProxyDor(uint64_t patchInde
 	}
 }
 
-EML2_NS::AbstractHdfProxy* PointsProperty::getValuesHdfProxyAndDatasetPathOfPatch(unsigned int patchIndex, std::string & datasetPath) const
+EML2_NS::AbstractHdfProxy* PointsProperty::getValuesHdfProxyAndDatasetPathOfPatch(uint64_t patchIndex, std::string & datasetPath) const
 {
 	if (patchIndex >= getPatchCount()) {
 		throw out_of_range("The values property patch is out of range");
