+++ src/resqml2_0_1/WellboreTrajectoryRepresentation.h
@@ -104,7 +104,7 @@ namespace RESQML2_0_1_NS
 
 		DLL_IMPORT_OR_EXPORT COMMON_NS::DataObjectReference getMdDatumDor() const final;
 
-		DLL_IMPORT_OR_EXPORT uint64_t getXyzPointCountOfPatch(unsigned int patchIndex) const final;
+		DLL_IMPORT_OR_EXPORT uint64_t getXyzPointCountOfPatch(uint64_t patchIndex) const final;
 
 		/**
 		 * @copybrief resqml2::AbstractRepresentation::getXyzPointsOfPatch
@@ -114,7 +114,7 @@ namespace RESQML2_0_1_NS
 		 * 
 		 * @copydetails resqml2::AbstractRepresentation::getXyzPointsOfPatch
 		 */
-		DLL_IMPORT_OR_EXPORT void getXyzPointsOfPatch(unsigned int patchIndex, double * xyzPoints) const final;
+		DLL_IMPORT_OR_EXPORT void getXyzPointsOfPatch(uint64_t patchIndex, double * xyzPoints) const final;
 
 		DLL_IMPORT_OR_EXPORT bool hasMdValues() const final;
