+++ src/resqml2_2/IjkGridLatticeRepresentation.cpp
@@ -383,7 +383,7 @@ void IjkGridLatticeRepresentation::setGeometryAsCoordinateLineNodes(
 }
 
 void IjkGridLatticeRepresentation::addSeismic3dCoordinatesToPatch(
-	unsigned int patchIndex,
+	uint64_t patchIndex,
 	double startInline, double incrInline, unsigned int countInline,
 	double startCrossline, double incrCrossline, unsigned int countCrossline,
 	unsigned int countSample, RESQML2_NS::AbstractRepresentation * seismicSupport)
